+++ src/main/java/org/braekpo1nt/mctmanager/ui/topbar/BattleTopbar.java
@@ -5,6 +5,9 @@ import com.google.common.base.Preconditions;
 import lombok.Data;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
+import org.braekpo1nt.mctmanager.ui.topbar.components.KillDeathComponent;
+import org.braekpo1nt.mctmanager.ui.topbar.components.TeamComponent;
+import org.braekpo1nt.mctmanager.ui.topbar.components.VersusComponent;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
